Rename 26.1 structured component folder
